+++ packages/store-engine-fs/jasmine.json
@@ -1,0 +1,6 @@
+{
+  "random": true,
+  "spec_dir": "src",
+  "spec_files": ["**/*.test.ts"],
+  "stopSpecOnExpectationFailure": true
+}
