+++ packages/store-engine-fs/package.json
@@ -1,13 +1,15 @@
 {
   "dependencies": {
     "@types/node": "~10",
-    "@wireapp/store-engine": "2.1.2"
+    "@wireapp/store-engine": "2.1.2",
+    "fs-extra": "8.0.1"
   },
   "devDependencies": {
     "@types/jasmine": "3.3.12",
     "jasmine": "3.4.0",
+    "nyc": "14.1.1",
     "rimraf": "2.6.3",
-    "ts-node": "8.1.0",
+    "ts-node": "8.2.0",
     "typescript": "3.4.5"
   },
   "description": "",
@@ -16,13 +18,13 @@
   ],
   "license": "GPL-3.0",
   "main": "./dist/index.js",
-  "name": "@wireapp/store-engine-sqleet",
+  "name": "@wireapp/store-engine-fs",
   "repository": "https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine-fs",
   "scripts": {
     "build": "tsc",
-    "clean": "rimraf dist",
+    "clean": "rimraf .nyc_output coverage dist",
     "dist": "yarn clean && yarn build",
-    "test": "karma start",
+    "test": "nyc jasmine --config=jasmine.json",
     "test:project": "yarn dist && yarn test"
   },
   "version": "0.0.0"
