+++ packages/store-engine-fs/src/index.ts
@@ -29,6 +29,7 @@ import * as fs from 'fs-extra';
 import * as path from 'path';
 
 export class FileEngine implements CRUDEngine {
+  [index: string]: any;
   public storeName = '';
   public options: {fileExtension: string} = {
     fileExtension: '.dat',
@@ -41,7 +42,7 @@ export class FileEngine implements CRUDEngine {
   public async isSupported(): Promise<void> {
     const isNode = typeof window === 'undefined';
 
-    if (isNode) {
+    if (!isNode) {
       const message = `Node.js File System Module is not available on your platform.`;
       throw new UnsupportedError(message);
     }
