+++ packages/store-engine/src/main/engine/FileEngine.test.node.ts
@@ -31,8 +31,6 @@ import {readAllSpec} from '../test/readAllSpec';
 import {readSpec} from '../test/readSpec';
 import {updateOrCreateSpec} from '../test/updateOrCreateSpec';
 import {updateSpec} from '../test/updateSpec';
-import {CRUDEngine} from './CRUDEngine';
-import {PathValidationError} from './error';
 
 const BASE_DIRECTORY = path.join(process.cwd(), '.tmp');
 const STORE_NAME = 'the-simpsons';
@@ -58,95 +56,6 @@ afterEach(done =>
     .then(done)
     .catch(done.fail)
 );
-
-describe('enforcePathRestrictions', () => {
-  const enforcePathRestrictions = (givenTrustedRoot: string, givenPath: string) => () =>
-    FileEngine.enforcePathRestrictions(givenTrustedRoot, givenPath);
-  const expectedError = PathValidationError;
-  const unixFolder = '/home/marge/test/';
-  const windowsFolder = 'C:\\Users\\bart\\Documents\\Database\\';
-
-  it('allows dots inside of primary keys.', () => {
-    const tableName = 'amplify';
-    const primaryKey = 'z.storage.StorageKey.EVENT.LAST_DATE';
-
-    FileEngine.path = path.posix;
-    const actual = FileEngine.enforcePathRestrictions(FileEngine.path.join(unixFolder, tableName), primaryKey);
-    expect(actual).toBeDefined();
-  });
-
-  it('allows slashes inside of primary keys as long as they do not navigate outside of the base folder.', () => {
-    FileEngine.path = path.posix;
-    expect(FileEngine.enforcePathRestrictions(unixFolder, 'users/..')).toBeDefined();
-    expect(FileEngine.enforcePathRestrictions(unixFolder, 'users/../')).toBeDefined();
-    expect(FileEngine.enforcePathRestrictions(unixFolder, 'users/../sandbox')).toBeDefined();
-    expect(FileEngine.enforcePathRestrictions(unixFolder, 'users/me')).toBeDefined();
-    expect(FileEngine.enforcePathRestrictions(unixFolder, 'a/b/c/d/e/f/g/../../../../ok')).toBeDefined();
-    expect(FileEngine.enforcePathRestrictions(unixFolder, 'a/b/c/../../../')).toBeDefined();
-    expect(enforcePathRestrictions(unixFolder, 'a/b/c/../../../../')).toThrowError(expectedError);
-  });
-
-  it('allows empty strings.', () => {
-    const tableName = 'amplify';
-    const primaryKey = '';
-
-    FileEngine.path = path.posix;
-    const actual = FileEngine.enforcePathRestrictions(FileEngine.path.join(unixFolder, tableName), primaryKey);
-    expect(actual).toBeDefined();
-  });
-
-  it('throws errors on path traversals.', () => {
-    FileEngine.path = path.win32;
-    expect(enforcePathRestrictions(windowsFolder, 'malicious\\..\\..\\test\\..\\..')).toThrowError(expectedError);
-    expect(enforcePathRestrictions(windowsFolder, '\\malicious\\..\\\\..entry\\..\\..')).toThrowError(expectedError);
-    expect(enforcePathRestrictions(windowsFolder, 'malicious\\..\\entry\\..\\..')).toThrowError(expectedError);
-    expect(enforcePathRestrictions(windowsFolder, '\\\\server\\..\\..\\..')).toThrowError(expectedError);
-    expect(enforcePathRestrictions(windowsFolder, 'malicious\\..\\..\\entry\\..\\')).toThrowError(expectedError);
-    expect(enforcePathRestrictions(windowsFolder, '..\\etc')).toThrowError(expectedError);
-
-    FileEngine.path = path.posix;
-    expect(enforcePathRestrictions(unixFolder, '../etc')).toThrowError(expectedError);
-    expect(enforcePathRestrictions(unixFolder, '/malicious/../../../entry/../test')).toThrowError(expectedError);
-    expect(enforcePathRestrictions(unixFolder, 'malicious/../../../entry/..')).toThrowError(expectedError);
-    expect(enforcePathRestrictions(unixFolder, 'documents/../../../../../etc/hosts')).toThrowError(expectedError);
-    expect(enforcePathRestrictions(unixFolder, 'malicious/../../../entry/../')).toThrowError(expectedError);
-    expect(enforcePathRestrictions(unixFolder, '../etc')).toThrowError(expectedError);
-    expect(enforcePathRestrictions(unixFolder, 'users/../../tigris')).toThrowError(expectedError);
-    expect(enforcePathRestrictions(unixFolder, 'users/../tigris/../../')).toThrowError(expectedError);
-  });
-
-  it('throws errors when attempting to use the root folder as a trusted root.', () => {
-    FileEngine.path = path.posix;
-    expect(enforcePathRestrictions('/', 'etc/hosts')).toThrowError(expectedError);
-
-    FileEngine.path = path.win32;
-    expect(enforcePathRestrictions('C:/', '\\Windows\\System32\\drivers\\etc\\hosts')).toThrowError(expectedError);
-  });
-
-  it('is applied to all store operations.', async done => {
-    const functionNames = [
-      'append',
-      'create',
-      'delete',
-      'deleteAll',
-      'read',
-      'readAll',
-      'readAllPrimaryKeys',
-      'update',
-      'updateOrCreate',
-    ];
-
-    for (const operation of functionNames) {
-      try {
-        await (engine as CRUDEngine)[operation]('../etc', 'primary-key', {});
-        done.fail();
-      } catch (error) {
-        expect(error instanceof expectedError).toBe(true);
-      }
-    }
-    done();
-  });
-});
 
 describe('init', () => {
   it('resolves with the directory to which the records will be saved.', async () => {
